moving totalEthPrice to hook
